Add check for query parameters on regex
It was found that redirects did not get mapped as expected when the
request URIs included query parameters. At the same time, the Android
application relies on redirects to REX not occuring when a URI has a
query parameter of `?minimal=true`. This update to the regex match
string should meet both of these needs.